requirements.txt
# Core data science stack
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
scipy>=1.11.0
statsmodels>=0.14.0

# Deep learning
tensorflow>=2.13.0

# Visualisation
matplotlib>=3.7.0
seaborn>=0.12.0

# IO & configuration
openpyxl>=3.1.0
pyyaml>=6.0

# Tooling & testing
pytest>=7.4.0
pytest-cov>=4.1.0
joblib>=1.3.0
